+++ R/performNormalization.R
@@ -2,7 +2,9 @@
 #' 
 #' This function allows users to normalize the enrichment calculations 
 #' by accounting for single-cell dropout and producing positive 
-#' values for downstream differential enrichment analyses. A positive range
+#' values for downstream differential enrichment analyses. Default calculation 
+#' uses will scale the enrichment values by the number of genes present from 
+#' the gene set and then use a natural log transformation. A positive range
 #' values is useful for several downstream analyses, like differential 
 #' evaluation for log2-fold change, but will alter the original 
 #' enrichment values.
@@ -39,17 +41,13 @@
 #' @export
 #' @return Single-cell object or matrix of normalized enrichment scores
 
-
-
-
-
 performNormalization <- function(sc.data,
-                                        enrichment.data = NULL,
-                                        assay = "escape",
-                                        gene.sets = NULL,
-                                        make.positive = FALSE,
-                                        scale.factor = NULL,
-                                        groups = NULL) {
+                                 enrichment.data = NULL,
+                                 assay = "escape",
+                                 gene.sets = NULL,
+                                 make.positive = FALSE,
+                                 scale.factor = NULL,
+                                 groups = NULL) {
   if(!is.null(assay)) {
     if(is_seurat_object(sc.data)) {
       assay.present <- assay %in% Assays(sc.data)
@@ -79,8 +77,7 @@ performNormalization <- function(sc.data,
   #Isolating the number of genes per cell expressed
   if(is.null(groups)){
     chunks <- dim(enriched)[[1]]
-  }
-  else{
+  } else{
     chunks <- min(groups, dim(enriched)[[1]])
   }
   
@@ -94,8 +91,7 @@ performNormalization <- function(sc.data,
     })
     egc.sizes <- split_rows(do.call(cbind,egc.sizes), chunk.size=chunks)
     rm(cnts)
-  }
-  else{
+  } else{
     egc.sizes <- split_vector(scale.factor, chunk.size=chunks)
   }
   enriched <- split_rows(enriched, chunk.size=chunks)
@@ -112,6 +108,13 @@ performNormalization <- function(sc.data,
       x+max(0, -min(x))
     })
   }
+  #Default Scaling using natural log
+  if(is.null(scale.factor)) {
+    enriched <- suppressWarnings(ifelse(enriched >= 0, 
+                                 log1p(enriched + 1e-6), 
+                                 -log1p(abs(enriched) + 1e-6)))
+  }
+  
   if(is_seurat_or_se_object(sc.data)) {
     if(is.null(assay)) {
       assay <- "escape"